fix: my previous change about ambiguity was totally wrong, I forgot ambiguity cannot be determined just by comparing the predicate size
Two rules can have the same priority and predicate size and still be unambiguous because their predicates are mutually exclusive. For example:

10: thing == 'foo' => bar = baz
10: thing == 'oof' => bar = zab

To know at the time of add(rules:) whether or not these are ambiguous would be difficult if not impossible (it seems like the enum exhaustiveness problem but it's possibly also the halting problem, and I don't really want to figure out which because either way it's not something I want to implement or support).

So now, when you ask a question, you'll get the rules that were found to be ambiguous during the processing of the answer to the question.

When there are ambiguous rules, one will be chosen and you should consider it undefined as to which candidate rule was chosen to provide the answer. You can log the rules that were found to be ambiguous to aid you in making unambiguous rules.